obsidian_agent/requirements.txt
openai
pyperclip
pyyaml 